backend/requirements.txt
acres==0.2.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
coverage==7.6.10
cryptography==44.0.0
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pluggy==1.5.0
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.3
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.2.0
pyserial==3.5
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
selenium==4.28.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
